% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handsontable4r-package.R
\name{handsontable4r-exports}
\alias{handsontable4r-exports}
\alias{\%>\%}
\title{handsontable4r exported operators}
\description{
The following functions are imported and then re-exported
from the handsontable4r package to enable use of the magrittr
pipe operator with no additional library calls
}
