% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handsontable4r-package.R
\docType{package}
\name{handsontable4r-package}
\alias{handsontable4r-package}
\title{handsontable4r}
\description{
R interface for creating tables using Handsontable v8, url{http://http://handsontable.com/}
}
