% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handsontable4r.R
\name{handsontable4r}
\alias{handsontable4r}
\title{Handsontable widget}
\usage{
handsontable4r(
  data,
  width = NULL,
  height = NULL,
  readonly = TRUE,
  rownames = NULL,
  rowHeaderWidth = NULL,
  digits = 3
)
}
\arguments{
\item{data}{data frame}

\item{width}{the width of widget}

\item{height}{the height of widget}

\item{readonly}{if the entire table is readonly}

\item{rownames}{the rownames}

\item{rowHeaderWidth}{the width of the header}

\item{digits}{the digits}
}
\description{
Create table
}
