% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handsontable4r.R
\name{handsontable4rOutput}
\alias{handsontable4rOutput}
\title{Handsontable widget}
\usage{
handsontable4rOutput(outputId, width = "100\%", height = "100\%")
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{must be a valid CSS unit in pixels
or a number, which will be coerced to a string and have \code{"px"} appended.}
}
\description{
Shiny bindings for handsontable4r
}
\seealso{
\code{\link{renderHandsontable4r}}
}
