% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handsontable4r.R
\name{hot_cell}
\alias{hot_cell}
\title{Handsontable widget}
\usage{
hot_cell(hot, row, col, comment = NULL, readOnly = NULL)
}
\arguments{
\item{hot}{handsontable4r object}

\item{row}{numeric row index}

\item{col}{column name or index}

\item{comment}{character comment to add to cell}

\item{readOnly}{logical making the cell read-only}
}
\description{
Configure single cell.  See
\href{http://handsontable.com}{Handsontable.js} for details.
}
\seealso{
\code{\link{hot_cols}}, \code{\link{hot_rows}}
}
