% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handsontable4r.R
\name{hot_row}
\alias{hot_row}
\title{Handsontable widget}
\usage{
hot_row(hot, row, readOnly = NULL)
}
\arguments{
\item{hot}{handsontable4r object}

\item{row}{numeric vector of row indexes}

\item{readOnly}{logical making the row(s) read-only}
}
\description{
Configure properties of all cells in a given row(s).
Note that hot_row is not to be confused with \code{\link{hot_rows}}.  See
\href{http://handsontable.com}{Handsontable.js} for details.
}
\seealso{
\code{\link{hot_cols}}, \code{\link{hot_cell}}, \code{\link{hot_rows}}
}
