% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handsontable4r.R
\name{hot_rows}
\alias{hot_rows}
\title{Handsontable widget}
\usage{
hot_rows(hot, rowHeights = NULL, fixedRowsTop = NULL)
}
\arguments{
\item{hot}{handsontable4r object}

\item{rowHeights}{a scalar or numeric vector of row heights}

\item{fixedRowsTop}{a scaler indicating the number of rows to
freeze on the top}
}
\description{
Configure row settings that pertain to the entire table.
Note that hot_rows is not to be confused with \code{\link{hot_row}}. See
\href{http://handsontable.com}{Handsontable.js} for details.
}
\seealso{
\code{\link{hot_cols}}, \code{\link{hot_cell}}
}
