% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handsontable4r.R
\name{renderHandsontable4r}
\alias{renderHandsontable4r}
\title{Handsontable widget}
\usage{
renderHandsontable4r(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{an expression that generates an handsontable4r}

\item{env}{the environment in which to evaluate \code{expr}.}

\item{quoted}{is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Shiny bindings for handsontable4r
}
\seealso{
\code{\link{Handsontable4rOutput}}
}
